requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
asttokens==3.0.0
attrs==25.4.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.61.0
frozenlist==1.8.0
graphviz==0.21
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.1.7
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
parso==0.8.5
pefile==2024.8.26
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.1
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.2
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
pyzmq==27.1.0
rsa==4.9.1
scipy==1.16.3
setuptools==82.0.0
six==1.17.0
stack-data==0.6.3
Telethon==1.42.0
tornado==6.5.2
traitlets==5.14.3
wcwidth==0.2.14
yarl==1.22.0
